+++ components/icon/stories/icon.story.css.js
@@ -6,7 +6,7 @@ const css = `.container {
 }
 
 .container > figure {
-	box-shadow: 1px 1px 3px 1px #ddd;
+	box-shadow: 0px 0px 3px 1px #12131411;
 	margin: 0.5rem;
 	width: 5rem;
 	height: 5rem;
@@ -14,20 +14,14 @@ const css = `.container {
 	text-align: center;
 	box-sizing: border-box;
 	position: relative;
-	background-color: #eee;
 	border-radius: 4px;
 	overflow: hidden;
+	color: #eee;
 	fill: #eee;
 }
 
-.container > figure:nth-child(5n+1)  { background-color: #FA97AA; }
-.container > figure:nth-child(5n+2)  { background-color: #F876C2; }
-.container > figure:nth-child(5n+3)  { background-color: #A93CF8; }
-.container > figure:nth-child(5n+4)  { background-color: #8D9BFA; }
-.container > figure:nth-child(5n+5)  { background-color: #F75CDA; }
-
 .container > figure > vwc-icon {
-	box-shadow: 3px 3px 3px rgba(0,0,0,0);
+	color: #121314;
 }
 
 .container > figure > figcaption {
@@ -41,8 +35,8 @@ const css = `.container {
 	width: 100%;
 	box-sizing: border-box;
 	font-weight: bold;
-	color: #eee;
-	background-color: #00000025;
+	color: #121314;
+	background-color: #12131422;
 	padding: 0.2rem;
 }`;