+++ CHANGELOG.md
@@ -3,6 +3,22 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+# 0.4.0 (2020-08-18)
+
+**Note:** Version bump only for package @vivid/root
+
+
+
+
+
+# 0.4.0 (2020-08-18)
+
+**Note:** Version bump only for package @vivid/root
+
+
+
+
+
 # 0.3.0 (2020-08-03)
 
 **Note:** Version bump only for package @vivid/root